+++ src/slice-chop.c
@@ -191,7 +191,17 @@ static SEXP chop(SEXP x, SEXP indices, struct vctrs_chop_info info) {
       ++(*info.p_index);
     }
 
-    SEXP elt = PROTECT(vec_slice_base(info.proxy_info.type, proxy, info.index));
+    // Always materialize ALTREP vectors when chopping to avoid inefficiently
+    // creating a large amount of small ALTREP objects that are used downstream.
+    // This is a heuristic and we should also be on the lookout for cases where
+    // we chop to create a small amount of large ALTREP objects that are
+    // quickly discarded (#1450).
+    SEXP elt = PROTECT(vec_slice_base(
+      info.proxy_info.type,
+      proxy,
+      info.index,
+      VCTRS_MATERIALIZE_true
+    ));
 
     if (names != R_NilValue) {
       SEXP elt_names = PROTECT(slice_names(names, info.index));